Chats order behaviour when receiving and sending message
chats will now move their position in the list based on when they send or receive a message. The chat gets moved to the top when a message is received or a message is sent. Chats are now also sorted based on when the chat was last altered using .updatedAt property.