Latest Message bug fix for receiving message and updated latestMessage
When a new message is received for a chat and the user is not on that chat. The chat will be moved to the top of the chats list however. The latestMessage field will not update because of the two setChats updates to the chats state. This was fixed by using newChats variable to do operations then assigning chats once done.